+++ lib/pages/settings/settings_card.dart
@@ -29,12 +29,7 @@ class SettingsCard extends StatelessWidget {
         leading: icon,
         title: Text(
           title,
-          style: TextStyle(
-            color: colorConfig().entryTextColor,
-            fontFamily: 'Oswald',
-            fontSize: 22,
-            fontWeight: FontWeight.w300,
-          ),
+          style: settingsCardTextStyle(),
         ),
         onTap: onTap,
       ),